pyproject.toml
[tool.poetry]
name = "prediction_football"
version = "0.1.0"
description = "Football prediction package"
authors = ["David Anders <danders232@gmail.com>"]
readme = "README.md"
packages = [
    { include = "*", from = "src" },
    { include = "*", from = "config" }
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pandas = "^2.2.3"
scikit-learn = "^1.5.2"
fuzzywuzzy = "^0.18.0"
mlflow = "^2.17.2"
python-googlesearch = "^1.1.1"
xgboost = "^2.1.2"
selenium = "^4.26.1"
fake-useragent = "^1.5.1"
dynaconf = "^3.2.6"

[tool.poetry.group.tensorflow.dependencies]
tensorflow = {version = "^2.13.0"}
tensorflow-intel = {version = "^2.13.0", platform = "win32"}
tensorflow-cpu = [
    {version = "^2.13.0", platform = "linux", markers = "platform_machine!='arm64' and platform_machine!='aarch64'"},
    {version = "^2.13.0", platform = "darwin", markers = "platform_machine!='arm64' and platform_machine!='aarch64'"}
]
tensorflow-cpu-aws = {version = "^2.13.0", platform = "linux", markers = "platform_machine=='arm64' or platform_machine=='aarch64'"}
tensorflow-io-gcs-filesystem = [
    {version = ">=0.23.1", markers = "platform_machine!='arm64' or platform_system!='Darwin'"},
    {version = "<0.32.0", markers = "platform_system == 'Windows'"}
]

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
